[#150] REFACTOR: 확장성을 위한 매물 관련 data providers 생성 및 리팩터링
