+++ api/src/main/java/org/silsagusi/api/article/application/RegionService.java
@@ -1,0 +1,21 @@
+package org.silsagusi.api.article.application;
+
+import lombok.RequiredArgsConstructor;
+import org.silsagusi.api.article.application.dto.RegionResponse;
+import org.silsagusi.api.article.infrastructure.dataProvider.RegionDataProvider;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class RegionService {
+
+	private final RegionDataProvider regionDataProvider;
+
+	public List<RegionResponse> getRegions() {
+		return regionDataProvider.getRegions().stream()
+			.map(RegionResponse::of)
+			.toList();
+	}
+}
