+++ api/src/main/java/org/silsagusi/api/article/infrastructure/repository/ComplexRepository.java
@@ -1,4 +1,4 @@
-package org.silsagusi.api.article.infrastructure;
+package org.silsagusi.api.article.infrastructure.repository;
 
 import org.silsagusi.core.domain.article.Complex;
 import org.springframework.data.jpa.repository.JpaRepository;
