--- core/src/main/java/org/silsagusi/core/domain/article/projection/RealEstateTypeRatioProjection.java
@@ -1,7 +1,0 @@
-package org.silsagusi.core.domain.article.projection;
-
-public interface RealEstateTypeRatioProjection {
-	String getType();
-
-	Integer getCount();
-}
